+++ armada_command/armada_api.py
@@ -2,11 +2,13 @@ from __future__ import print_function
 
 import json
 import os
-import urllib
+import sys
+import traceback
 
 import requests
 
 from armada_command.consul.consul import consul_query
+from armada_utils import is_verbose, print_err, ship_name_to_ip
 
 
 def __are_we_in_armada_container():
@@ -22,12 +24,13 @@ def __get_armada_address(ship_name=None):
             if service['Service'] == 'armada':
                 return 'http://127.0.0.1:{}'.format(service['Port'])
     else:
+        ship_ip = ship_name_to_ip(ship_name)
         service_armada_dict = consul_query('catalog/service/armada')
         for service_armada in service_armada_dict:
-            if service_armada['Node'] in (ship_name, 'ship-' + ship_name) or service_armada['Address'] == ship_name:
-                return 'http://{0}:{1}'.format(service_armada['Address'], service_armada['ServicePort'])
+            if service_armada['Address'] == ship_ip:
+                return 'http://{0}:{1}'.format(ship_ip, service_armada['ServicePort'])
 
-    raise ValueError('Cannot find ship: {ship_name}.'.format(ship_name=ship_name))
+    raise ValueError('Cannot find ship: {0}.'.format(ship_name))
 
 
 def __exception_to_status(e):
@@ -39,16 +42,38 @@ def __exception_to_status(e):
 def get(api_function, arguments=None, ship_name=None):
     arguments = arguments or {}
     try:
-        result = requests.get(__get_armada_address(ship_name) + '/' + api_function + '?' + urllib.urlencode(arguments))
+        url = __get_armada_address(ship_name) + '/' + api_function
+        result = requests.get(url, params=arguments)
+        result.raise_for_status()
         return result.text
     except Exception as e:
+        if is_verbose():
+            traceback.print_exc()
         return __exception_to_status(e)
 
 
 def post(api_function, arguments=None, ship_name=None):
     arguments = arguments or {}
     try:
-        result = requests.post(__get_armada_address(ship_name) + '/' + api_function, json.dumps(arguments))
+        url = __get_armada_address(ship_name) + '/' + api_function
+        result = requests.post(url, json=arguments)
+        result.raise_for_status()
         return result.json()
     except Exception as e:
+        if is_verbose():
+            traceback.print_exc()
         return __exception_to_status(e)
+
+
+def print_result_from_armada_api(result):
+    if result['status'] == 'ok':
+        result_value = dict(result)
+        del result_value['status']
+        if result_value:
+            print(json.dumps(result_value))
+    else:
+        if result['status'] == 'error':
+            print_err(result.get('error'))
+        else:
+            print_err(result)
+        sys.exit(1)